requirements.txt
python-telegram-bot
google-generativeai>=0.5.0
flask